cleaned up old grpc code. started doc
